+++ site/src/components/ChildBounty/Detail/Meta/actions/hunter/acceptAndStart/index.js
@@ -1,6 +1,6 @@
 // TODO: uncompleted
 import { Button } from "@osn/common-ui";
-import { ButtonGroup } from "../styled";
+import { ButtonGroup } from "../../styled";
 
 export function useAcceptAndStart() {
   return (
